fix(api/identity): handle delete for non existent user

+++ src/apps/api/src/modules/identity/application/use-cases/DeleteUserByIdUseCase.ts
@@ -9,6 +9,13 @@ export class DeleteUserByIdUseCase {
   ) {}
 
   async execute(id: string) {
+
+    const user = await this.userRepository.getById(id)
+
+    if(!user) {
+      throw new Error('User not found')
+    }
+
     await this.userRepository.deleteById(id);
   }
 }
